Add files via upload
This is from my first attempt. The data analysis is by no means incredibly high level, but I believe that it does provide a strong summary of the health expenditures of different states, as well as important insights into the spending and funding differences across different states. 
Next steps: I hope to plot and / or visualise this data in a neat and effective way.